requirements.txt
python-dotenv>=0.10
requests>=2.21
instabot>=0.30
Pillow>=6.0